[#2527] Move strategy exception to a common file, unify imports
