+++ amy/emails/actions/exceptions.py
@@ -1,0 +1,2 @@
+class EmailStrategyException(Exception):
+    pass
